[#42309] Opt-in to Gitlab-runners on Kubernetes
This opts-in to use the new Gitlab-runners based on Kubernetes. The plan
is for the Kubernetes based runners to replace the old expensive
VM+docker based runners. `salt-automation` and `os2mo` have been using
kubernetes based runners for some weeks. After a few bugs have been
fixed, it is now time to try it on a more diverse set of jobs:
os2borgerpc-image is such a candidate. We do expect a small amount of
job failures to occur. Please report any job failures with a non-trivial
cause. Especially jobs that fail with `runner-failure`.

If this cause a larger issue, you can easily opt-out again by replacing
the `kubernetes`-tag with a `docker`-tag.

All Kubernetes runners are as capable as the old runners tagged with
`high-mem`.